+++ priv/repo/migrations/20180731145345_add_error_to_modlog_event_enum.exs
@@ -1,0 +1,8 @@
+defmodule Bolt.Repo.Migrations.AddErrorToModlogEventEnum do
+  use Ecto.Migration
+  @disable_ddl_transaction true
+
+  def up do
+    execute("ALTER TYPE modlog_event ADD VALUE 'ERROR';")
+  end
+end