+++ priv/repo/migrations/20180731150442_add_delete_invocation_to_accept_action_type_enum.exs
@@ -1,0 +1,8 @@
+defmodule Bolt.Repo.Migrations.AddDeleteInvocationToAcceptActionTypeEnum do
+  use Ecto.Migration
+  @disable_ddl_transaction true
+
+  def up do
+    execute("ALTER TYPE accept_action_type ADD VALUE 'delete_invocation';")
+  end
+end